services/file_watcher/requirements.txt
aiofiles~=24.1.0
nats-py
click~=8.1.7
pyyaml~=6.0.1
watchdog~=4.0.1
